Show message when loading save file for inspector
