+++ PalCalc.SaveReader/SaveFile/ISaveFile.cs
@@ -27,7 +27,7 @@ namespace PalCalc.SaveReader.SaveFile
         {
             CompressedSAV.WithDecompressedSave(FilePath, stream =>
             {
-                using (var archiveReader = new FArchiveReader(stream, PalWorldTypeHints.Hints, true))
+                using (var archiveReader = new FArchiveReader(stream, PalWorldTypeHints.Hints, false))
                     GvasFile.FromFArchive(archiveReader, visitors);
             });
         }
