+++ PalCalc.UI/View/LoadingSaveFileModal.xaml.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using System.Threading;
 using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Controls;
@@ -11,6 +12,7 @@ using System.Windows.Input;
 using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Shapes;
+using System.Windows.Threading;
 
 namespace PalCalc.UI.View
 {
@@ -19,9 +21,31 @@ namespace PalCalc.UI.View
     /// </summary>
     public partial class LoadingSaveFileModal : Window
     {
+        // https://stackoverflow.com/a/73181682
+        private static void AllowUIToUpdate()
+        {
+            DispatcherFrame frame = new();
+            // DispatcherPriority set to Input, the highest priority
+            Dispatcher.CurrentDispatcher.Invoke(DispatcherPriority.Input, new DispatcherOperationCallback(delegate (object parameter)
+            {
+                frame.Continue = false;
+                Thread.Sleep(20); // Stop all processes to make sure the UI update is perform
+                return null;
+            }), null);
+            Dispatcher.PushFrame(frame);
+            // DispatcherPriority set to Input, the highest priority
+            Application.Current.Dispatcher.Invoke(DispatcherPriority.Input, new Action(delegate { }));
+        }
+
         public LoadingSaveFileModal()
         {
             InitializeComponent();
+        }
+
+        public void ShowSync()
+        {
+            this.Show();
+            AllowUIToUpdate();
         }
     }
 }
